feat: :compile_auto_inspect and :auto_inspect options
If the first is false, inspect overrides will not be compiled. If
compiled and the second is false, inspect override will not be triggered
until it is set to true.

Would like to hook into inspect opts with `scribe: false` but opts are
always passed as a struct. Perhaps might use `pretty: false` instead.